+++ Assets/Prefabs/Modules/P_Module_Item_T2.prefab
@@ -54,10 +54,14 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5304199223535757148, guid: 4f979c97e22c1d441a5f5c8c657671e0, type: 3}
       propertyPath: m_Name
-      value: P_Upgrade_Card_Item_T2
+      value: P_Module_Item_T2
       objectReference: {fileID: 0}
     - target: {fileID: 5850868388823037643, guid: 4f979c97e22c1d441a5f5c8c657671e0, type: 3}
       propertyPath: cardTeir
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5850868388823037643, guid: 4f979c97e22c1d441a5f5c8c657671e0, type: 3}
+      propertyPath: cardTier
       value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
