+++ Assets/Prefabs/Modules/P_Module_Item_T3.prefab
@@ -74,9 +74,13 @@ PrefabInstance:
       propertyPath: cardTeir
       value: 2
       objectReference: {fileID: 0}
+    - target: {fileID: 1077538669313687144, guid: 1fc9c75036037e94cae04f2af9b72751, type: 3}
+      propertyPath: cardTier
+      value: 2
+      objectReference: {fileID: 0}
     - target: {fileID: 1610625579557318655, guid: 1fc9c75036037e94cae04f2af9b72751, type: 3}
       propertyPath: m_Name
-      value: P_Upgrade_Card_Item_T3
+      value: P_Module_Item_T3
       objectReference: {fileID: 0}
     - target: {fileID: 6830441041008247905, guid: 1fc9c75036037e94cae04f2af9b72751, type: 3}
       propertyPath: 'm_Materials.Array.data[0]'
