+++ Assets/Scripts/Data/Stats/Base/UpgradableStat.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b7bae2b1f79cf4b43b57ef4c006aa265
