+++ Assets/Scripts/Data/Stats/Enemies/RangedEnemyStatsSO.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 // TODO: Use a base class ideally but eh, screw it.
 
 [Serializable]
-public class RangedEnemyStats : ICloneable
+public class RangedEnemyStats : CoreStats, ICloneable
 {
     public float health = 0f;
-    
+
     public float damage = 0f;
     public float attackRange = 0f;
     public float attackSpeed = 0f;
@@ -19,11 +19,21 @@ public class RangedEnemyStats : ICloneable
     public float dodgeSpeed = 0f;
 
 
-    public object Clone()
+    public override object Clone()
     {
         return new RangedEnemyStats
         {
             health = health,
+
+            damage = damage,
+            attackRange = attackRange,
+            attackSpeed = attackSpeed,
+
+            baseSpeed = baseSpeed,
+            followSpeed = followSpeed,
+            fleeSpeed = fleeSpeed,
+            stuckSpeed = stuckSpeed,
+            dodgeSpeed = dodgeSpeed,
         };
     }
 
@@ -33,7 +43,7 @@ public class RangedEnemyStats : ICloneable
 public class RangedEnemyStatsSO : StatsCoreSO
 {
     // Please make sure the variables that you want to access are not able to be modified.
-    // Example below shows you one way to achive this.
+    // Example below shows you one way to achieve this.
 
     [SerializeField]
     private RangedEnemyStats stats;
