+++ Assets/Scripts/Data/Stats/Player/PlayerStatsSO.cs
@@ -2,18 +2,22 @@ using System;
 using UnityEngine;
 
 // TODO: Use a base class ideally but eh, screw it.
+// ? Do you mean for like health, move speed etc? because it does use a base class.
 
+/// <summary>
+/// Player stats class storing the players upgradable stats and constant stats.
+/// </summary>
 [Serializable]
 public class PlayerStats : CoreStats, ICloneable
 {
-    public UpgradablePlayerStat MaxHealthStat = new UpgradablePlayerStat(100f);
+    public UpgradableStat MaxHealthStat = new UpgradableStat(100f);
 
 
     public float RegenerationSpeed = 1f;
-    public UpgradablePlayerStat RegenerationAmountStat = new(5f);
+    public UpgradableStat RegenerationAmountStat = new(5f);
 
     public float WalkSpeed = 4f;
-    public UpgradablePlayerStat GroundRunSpeedStat = new(5.5f);
+    public UpgradableStat GroundRunSpeedStat = new(5.5f);
     public float AirSpeed = 3f; // TODO: should be a fraction from gound speed.
 
     // public float GroundAcceleration = 30f;
@@ -26,33 +30,33 @@ public class PlayerStats : CoreStats, ICloneable
     public float JumpForce = 9.2f;
 
 
-    public UpgradablePlayerStat SlideBoostPercentageStat = new(1.65f);
-    public UpgradablePlayerStat AirBoostPercentageStat = new(2.5f);
+    public UpgradableStat SlideBoostPercentageStat = new(1.65f);
+    public UpgradableStat AirBoostPercentageStat = new(2.5f);
 
     public float GroundFriction = 5f;
     public float AirFriction = 1;
 
 
-    public UpgradablePlayerStat ProjectileDamageStat = new(12f);
-    public UpgradablePlayerStat RechargeSecondsStat = new(0.3f);
-    public UpgradablePlayerStat ShotsPerFullChargeStat = new(10);
+    public UpgradableStat ProjectileDamageStat = new(12f);
+    public UpgradableStat RechargeSecondsStat = new(0.3f);
+    public UpgradableStat ShotsPerFullChargeStat = new(10);
     public float StandardSecondsPerShot = 0.4f;
     public float ChargedSecondsPerShot = 0.1f;
     public float DelayAfterFireBeforeRecharging = 0.4f;
-    public UpgradablePlayerStat OverheatForceCoolDownStat = new(3f);
+    public UpgradableStat OverheatForceCoolDownStat = new(3f);
 
 
-    public UpgradablePlayerStat MeleeAttackDelayStat = new(0.5f);
-    public UpgradablePlayerStat MeleeDamageStat = new(10f);
-    public UpgradablePlayerStat MeleeReachStat = new(1.5f);
-    public UpgradablePlayerStat MeleeStaggerTimeStat = new(0.4f);
+    public UpgradableStat MeleeAttackDelayStat = new(0.5f);
+    public UpgradableStat MeleeDamageStat = new(10f);
+    public UpgradableStat MeleeReachStat = new(1.5f);
+    public UpgradableStat MeleeStaggerTimeStat = new(0.4f);
 
-    public UpgradablePlayerStat BashForceStat = new(10f);
-    public UpgradablePlayerStat BashAttackDelayStat = new(0.5f);
+    public UpgradableStat BashForceStat = new(10f);
+    public UpgradableStat BashAttackDelayStat = new(0.5f);
 
-    protected override UpgradablePlayerStat[] GetAllUpgradableStats()
+    protected override UpgradableStat[] GetAllUpgradableStats()
     {
-        UpgradablePlayerStat[] upgradablePlayerStats =
+        UpgradableStat[] upgradablePlayerStats =
         {
             MaxHealthStat,
             RegenerationAmountStat,
@@ -78,13 +82,13 @@ public class PlayerStats : CoreStats, ICloneable
     {
         return new PlayerStats
         {
-            MaxHealthStat = (UpgradablePlayerStat)MaxHealthStat.Clone(),
+            MaxHealthStat = (UpgradableStat)MaxHealthStat.Clone(),
 
             RegenerationSpeed = RegenerationSpeed,
-            RegenerationAmountStat = (UpgradablePlayerStat)RegenerationAmountStat.Clone(),
+            RegenerationAmountStat = (UpgradableStat)RegenerationAmountStat.Clone(),
 
             WalkSpeed = WalkSpeed,
-            GroundRunSpeedStat = (UpgradablePlayerStat)GroundRunSpeedStat.Clone(),
+            GroundRunSpeedStat = (UpgradableStat)GroundRunSpeedStat.Clone(),
             AirSpeed = AirSpeed,
 
             GroundAccelerationPercentBase = GroundAccelerationPercentBase,
@@ -95,37 +99,40 @@ public class PlayerStats : CoreStats, ICloneable
             // SlideBoostForce = SlideBoostForce,
             // AirBoostForce = AirBoostForce,
 
-            SlideBoostPercentageStat = (UpgradablePlayerStat)SlideBoostPercentageStat.Clone(),
-            AirBoostPercentageStat = (UpgradablePlayerStat)AirBoostPercentageStat.Clone(),
+            SlideBoostPercentageStat = (UpgradableStat)SlideBoostPercentageStat.Clone(),
+            AirBoostPercentageStat = (UpgradableStat)AirBoostPercentageStat.Clone(),
 
-            ProjectileDamageStat = (UpgradablePlayerStat)ProjectileDamageStat.Clone(),
-            RechargeSecondsStat = (UpgradablePlayerStat)RechargeSecondsStat.Clone(),
-            ShotsPerFullChargeStat = (UpgradablePlayerStat)ShotsPerFullChargeStat.Clone(),
+            ProjectileDamageStat = (UpgradableStat)ProjectileDamageStat.Clone(),
+            RechargeSecondsStat = (UpgradableStat)RechargeSecondsStat.Clone(),
+            ShotsPerFullChargeStat = (UpgradableStat)ShotsPerFullChargeStat.Clone(),
             StandardSecondsPerShot = StandardSecondsPerShot,
             ChargedSecondsPerShot = ChargedSecondsPerShot,
             DelayAfterFireBeforeRecharging = DelayAfterFireBeforeRecharging,
-            OverheatForceCoolDownStat = (UpgradablePlayerStat)OverheatForceCoolDownStat.Clone(),
+            OverheatForceCoolDownStat = (UpgradableStat)OverheatForceCoolDownStat.Clone(),
             // ProjectileFireRate = ProjectileFireRate,
 
 
-            MeleeAttackDelayStat = (UpgradablePlayerStat)MeleeAttackDelayStat.Clone(),
-            MeleeDamageStat = (UpgradablePlayerStat)MeleeDamageStat.Clone(),
-            BashForceStat = (UpgradablePlayerStat)BashForceStat.Clone(),
-            BashAttackDelayStat = (UpgradablePlayerStat)BashAttackDelayStat.Clone(),
+            MeleeAttackDelayStat = (UpgradableStat)MeleeAttackDelayStat.Clone(),
+            MeleeDamageStat = (UpgradableStat)MeleeDamageStat.Clone(),
+            BashForceStat = (UpgradableStat)BashForceStat.Clone(),
+            BashAttackDelayStat = (UpgradableStat)BashAttackDelayStat.Clone(),
 
-            MeleeReachStat = (UpgradablePlayerStat)MeleeReachStat.Clone(),
-            MeleeStaggerTimeStat = (UpgradablePlayerStat)MeleeStaggerTimeStat.Clone(),
+            MeleeReachStat = (UpgradableStat)MeleeReachStat.Clone(),
+            MeleeStaggerTimeStat = (UpgradableStat)MeleeStaggerTimeStat.Clone(),
 
         };
     }
 
 }
 
+/// <summary>
+/// The player scriptable object storing the player stats.
+/// </summary>
 [CreateAssetMenu(menuName = "ScriptableObjects/Stats/PlayerStats", fileName = "SO_PlayerStats")]
 public class PlayerStatsSO : StatsCoreSO
 {
-    // Please make sure the variables that you want to access are not able to be modified.
-    // Example below shows you one way to achive this.
+    // Please make sure the variables that you want to access are not able to be modified directly.
+    // Example below shows you one way to achieve this.
 
     [SerializeField]
     private PlayerStats stats;
@@ -135,7 +142,7 @@ public class PlayerStatsSO : StatsCoreSO
         return stats.Clone();
     }
 
-    void OnValidate()
+    void OnValidate() // The work around to reset and set data correctly at unity validate.
     {
         stats.RefreshStats();
         stats.ResetAllChipStats(); // make sure there are no lingering data.
