+++ Assets/Scripts/Data/Stats/Player/PlayerStatsSetter.cs
@@ -1,5 +1,8 @@
 using UnityEngine;
 
+/// <summary>
+/// Reads the player stats from the stat manager or creates dummy class if there is no manager and sets it to the player.
+/// </summary>
 public class PlayerStatsSetter : MonoBehaviour
 {
     // Start is called once before the first execution of Update after the MonoBehaviour is created
@@ -25,11 +28,5 @@ public class PlayerStatsSetter : MonoBehaviour
         if (stats != null) GetComponent<Regeneration>().SetUpRegeneration(stats.RegenerationSpeed / stats.RegenerationAmountStat.GetCurrentValue(), 3f);
         GetComponent<PlayerCombat>().UpdateVariablesWithStats(stats);
         GetComponent<PlayerMovement>().UpdateVariablesWithStats(stats);
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
     }
 }
