+++ Assets/Scripts/Environment/Collectables/ModuleCollectables.meta
