+++ Assets/Scripts/Environment/Collectables/ModuleCollectables/ModuleCollectable.cs.meta
