--- Assets/Scripts/Environment/Collectables/UpgradeCardCollectables/CardCollectable.cs
@@ -1,17 +1,0 @@
-using UnityEngine;
-
-public class CardCollectable : CollectableBase
-{
-    public ModuleTier cardTeir = ModuleTier.Common;
-
-    protected override void CollectItem()
-    {
-        if (!CanPlayerCollect()) return;
-
-        if (Vector3.Distance(transform.position, targetTransform.position) > collectItemRange) return;
-
-        ModuleManager.Instance.CollectModule(cardTeir);
-        Destroy(gameObject);
-    }
-
-}
