+++ Assets/Scripts/Environment/DepotAnimation.meta
