+++ Assets/Scripts/Environment/Doors/Crushing.cs
@@ -1,24 +1,26 @@
 using System.Collections.Generic;
 using UnityEngine;
 
+/// <summary>
+/// Crush entities when they are added to the stored list twice.
+/// <br />This is directly tied to doors. // TODO: Remove the door link and have a separate script.
+/// </summary>
 public class Crushing : MonoBehaviour
 {
     private List<Collider> detectedColliders = new List<Collider>();
 
-    //private float damage = 99999f;
-
     private bool isActive = true;
 
-    private Door door;
+    private Door door; // TODO: REMOVE.
 
     void Start()
     {
-        door = GetComponent<Door>();
+        door = GetComponent<Door>(); // TODO: REMOVE.
     }
 
     void Update()
     {
-        if (door.IsDoorOpen())
+        if (door.IsDoorOpen()) // TODO: REMOVE.
         {
             isActive = false;
         }
@@ -28,15 +30,16 @@ public class Crushing : MonoBehaviour
         }
     }
 
+    // If this does not work, then replace collider with health.
     public void AddCollider(Collider other)
     {
         if (!isActive) return;
 
-        // print(other.name);
+
         if (detectedColliders.Contains(other))
         {
-            //other.GetComponent<Health>()?.AddToHealth(-damage);
             other.GetComponent<Health>()?.InstantKill();
+
             detectedColliders.Remove(other);
         }
         else
