+++ Assets/Scripts/Environment/Doors/CrushingDetector.cs
@@ -8,12 +8,24 @@ public class CrushingDetector : MonoBehaviour
     void OnTriggerEnter(Collider other)
     {
         if (other.GetComponent<Health>() != null)
+        {
             crushing.AddCollider(other);
+        }
+        else if (other.GetComponentInParent<Health>() != null && other.GetComponentInParent<Collider>() != null)
+        {
+            crushing.AddCollider(other.GetComponentInParent<Collider>());
+        }
     }
 
     void OnTriggerExit(Collider other)
     {
         if (other.GetComponent<Health>() != null)
+        {
             crushing.RemoveCollider(other);
+        }
+        else if (other.GetComponentInParent<Health>() != null && other.GetComponentInParent<Collider>() != null)
+        {
+            crushing.RemoveCollider(other.GetComponentInParent<Collider>());
+        }
     }
 }
